requirements.txt
Aliyun==0.1.1
aliyun-python-sdk==2.2.0
aliyun-python-sdk-cms==3.2.7
aliyun-python-sdk-core==2.0.36
aliyun-python-sdk-ecs==2.1.0
aliyuncli==2.1.2
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
blessings==1.6
bpython==0.15
certifi==2016.2.28
cffi==1.7.0
colorama==0.3.3
cryptography==1.4
curtsies==0.2.6
Django==1.8.2
django-bootstrap-toolkit==2.15.0
djangorestframework==3.4.1
enum34==1.1.6
futures==3.0.5
greenlet==0.4.10
httplib2==0.9.2
idna==2.1
ipaddress==1.0.16
Jinja2==2.8
jmespath==0.7.1
M2Crypto==0.25.1
MarkupSafe==0.23
msgpack-python==0.4.7
MySQL-python==1.2.3
ndg-httpsclient==0.4.2
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
Pygments==2.1.3
pyOpenSSL==16.0.0
PyYAML==3.11
pyzmq==15.3.0
requests==2.10.0
salt==2015.5.8
singledispatch==3.4.0.3
six==1.10.0
tornado==4.4.1
typing==3.5.2.2
wcwidth==0.1.7
